SSD_WEEK5/backend/requirements.txt
Flask
Flask_SQLAlchemy
python-dotenv
psycopg2-binary
PyJWT
